Conflicts:
	Book/WebContent/WEB-INF/views/horrorBoard.jsp
	Book/WebContent/WEB-INF/views/join_success.jsp
	Book/WebContent/WEB-INF/views/main.jsp
	Book/WebContent/WEB-INF/views/read.jsp
	Book/WebContent/WEB-INF/views/write_form.jsp
	Book/WebContent/WEB-INF/views/write_result.jsp
	Book/src/controller/BookController.java
	Book/src/mapper/CommentMapper.java
	Book/src/mapper/comment_mapper.xml
	Book/src/repository/CommentDAO.java
	Book/src/service/BookBoardService.java
	Book/target/m2e-wtp/web-resources/resource/style.css

+++ Book/src/service/AdminService.java
@@ -1,9 +1,12 @@
 package service;
 
 import java.util.List;
+import java.util.Random;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import repository.AdminDAO;
 import repository.MemberDAO;
@@ -89,7 +92,13 @@ public class AdminService {
 	}
 	
 	// kick
-	public boolean kick(String id){
+	@Transactional(propagation = Propagation.REQUIRED, rollbackFor={Exception.class})
+	public boolean kick(String id) throws Exception{
+		MemberVO member = mdao.selectDeleteMem(id);
+		mdao.insertDeleteMem(member);
+		if(new Random().nextBoolean()){
+			throw new Exception("에러");
+		}
 		if(mdao.deleteMem(id)>0){
 			return true;
 		}else{
